+++ lib/src/layer.dart
@@ -43,6 +43,10 @@ abstract class Layer {
   /// Each type is associated with a concrete implementation of [Layer].
   LayerType type;
 
+  /// The "Class" specified in Tiled, introduced in Tiled 1.9 to support
+  /// custom types on any object. This is NOT the same as [type]
+  String? class_;
+
   /// Horizontal layer offset in tiles. Always 0.
   int x;
 
@@ -86,6 +90,7 @@ abstract class Layer {
     this.id,
     required this.name,
     required this.type,
+    this.class_,
     this.x = 0,
     this.y = 0,
     this.offsetX = 0,
@@ -108,6 +113,7 @@ abstract class Layer {
 
     final id = parser.getIntOrNull('id');
     final name = parser.getString('name', defaults: '');
+    final class_ = parser.getStringOrNull('class');
     final x = parser.getInt('x', defaults: 0);
     final y = parser.getInt('y', defaults: 0);
     final offsetX = parser.getDouble('offsetx', defaults: 0);
@@ -144,6 +150,7 @@ abstract class Layer {
         layer = TileLayer(
           id: id,
           name: name,
+          class_: class_,
           x: x,
           y: y,
           offsetX: offsetX,
@@ -174,6 +181,7 @@ abstract class Layer {
         layer = ObjectGroup(
           id: id,
           name: name,
+          class_: class_,
           x: x,
           y: y,
           offsetX: offsetX,
@@ -194,15 +202,20 @@ abstract class Layer {
       case LayerType.imageLayer:
         final transparentColor = parser.getStringOrNull('transparentcolor');
         final image = parser.getSingleChildAs('image', TiledImage.parse);
+        final repeatX = parser.getBool('repeatx', defaults: false);
+        final repeatY = parser.getBool('repeaty', defaults: false);
         layer = ImageLayer(
           id: id,
           name: name,
+          class_: class_,
           x: x,
           y: y,
           offsetX: offsetX,
           offsetY: offsetY,
           parallaxX: parallaxX,
           parallaxY: parallaxY,
+          repeatX: repeatX,
+          repeatY: repeatY,
           startX: startX,
           startY: startY,
           tintColor: tintColor,
@@ -218,6 +231,7 @@ abstract class Layer {
         layer = Group(
           id: id,
           name: name,
+          class_: class_,
           x: x,
           y: y,
           offsetX: offsetX,
@@ -242,10 +256,13 @@ abstract class Layer {
     return parser.formatSpecificParsing(
       (json) => json.getChildrenAs('layers', Layer.parse),
       (xml) {
-        return xml.getChildrenAs('layer', Layer.parse) +
-            xml.getChildrenAs('objectgroup', Layer.parse) +
-            xml.getChildrenAs('imagelayer', Layer.parse) +
-            xml.getChildrenAs('group', Layer.parse);
+        // It's very important not change the order of the layers
+        // during parsing!
+        // Order in the map determines rendering order.
+        final xmlLayers = xml.getChildrenWithNames(
+          {'layer', 'objectgroup', 'imagelayer', 'group'},
+        );
+        return xmlLayers.map(Layer.parse).toList();
       },
     );
   }
@@ -348,6 +365,7 @@ class TileLayer extends Layer {
   TileLayer({
     int? id,
     required String name,
+    String? class_,
     int x = 0,
     int y = 0,
     double offsetX = 0,
@@ -371,6 +389,7 @@ class TileLayer extends Layer {
           id: id,
           name: name,
           type: LayerType.tileLayer,
+          class_: class_,
           x: x,
           y: y,
           offsetX: offsetX,
@@ -410,6 +429,7 @@ class ObjectGroup extends Layer {
   ObjectGroup({
     int? id,
     required String name,
+    String? class_,
     int x = 0,
     int y = 0,
     double offsetX = 0,
@@ -429,6 +449,7 @@ class ObjectGroup extends Layer {
           id: id,
           name: name,
           type: LayerType.objectGroup,
+          class_: class_,
           x: x,
           y: y,
           offsetX: offsetX,
@@ -451,9 +472,16 @@ class ImageLayer extends Layer {
   /// Hex-formatted color (#RRGGBB) (optional).
   String? transparentColor;
 
+  /// Whether or not to repeat the image on the X-axis
+  bool repeatX;
+
+  /// Whether or not to repeat the image on the Y-axis
+  bool repeatY;
+
   ImageLayer({
     int? id,
     required String name,
+    String? class_,
     int x = 0,
     int y = 0,
     double offsetX = 0,
@@ -467,11 +495,14 @@ class ImageLayer extends Layer {
     bool visible = true,
     List<Property> properties = const [],
     required this.image,
+    required this.repeatX,
+    required this.repeatY,
     this.transparentColor,
   }) : super(
           id: id,
           name: name,
           type: LayerType.imageLayer,
+          class_: class_,
           x: x,
           y: y,
           offsetX: offsetX,
@@ -494,6 +525,7 @@ class Group extends Layer {
   Group({
     int? id,
     required String name,
+    String? class_,
     int x = 0,
     int y = 0,
     double offsetX = 0,
@@ -511,6 +543,7 @@ class Group extends Layer {
           id: id,
           name: name,
           type: LayerType.imageLayer,
+          class_: class_,
           x: x,
           y: y,
           offsetX: offsetX,